+++ app/src/main/java/io/github/wulkanowy/data/db/entities/Grade.kt
@@ -34,7 +34,7 @@ data class Grade(
 
     val weight: String,
 
-    val weightValue: Int,
+    val weightValue: Double,
 
     val date: LocalDate,
